+++ src/main/java/com/khay/gestiondestock/services/impl/FournisseurServiceImpl.java
@@ -1,8 +1,12 @@
 package com.khay.gestiondestock.services.impl;
 
 import com.khay.gestiondestock.dto.FournisseurDto;
+import com.khay.gestiondestock.exception.EntityNotFoundException;
 import com.khay.gestiondestock.exception.ErrorCodes;
 import com.khay.gestiondestock.exception.InvalidEntityException;
+import com.khay.gestiondestock.exception.InvalidOperationException;
+import com.khay.gestiondestock.model.CommandeFournisseur;
+import com.khay.gestiondestock.repository.CommandeFournisseurRepository;
 import com.khay.gestiondestock.repository.FournisseurRepository;
 import com.khay.gestiondestock.services.FournisseurService;
 import com.khay.gestiondestock.validator.FournisseurValidator;
@@ -11,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -18,9 +23,12 @@ public class FournisseurServiceImpl implements FournisseurService {
 
     private FournisseurRepository fournisseurRepository;
 
+    private CommandeFournisseurRepository commandeFournisseurRepository;
+
     @Autowired
-    public FournisseurServiceImpl(FournisseurRepository fournisseurRepository) {
+    public FournisseurServiceImpl(FournisseurRepository fournisseurRepository, CommandeFournisseurRepository commandeFournisseurRepository) {
         this.fournisseurRepository = fournisseurRepository;
+        this.commandeFournisseurRepository = commandeFournisseurRepository;
     }
 
 
@@ -28,29 +36,49 @@ public class FournisseurServiceImpl implements FournisseurService {
     public FournisseurDto save(FournisseurDto fournisseur) {
         List<String> errors = FournisseurValidator.validate(fournisseur);
         if (!errors.isEmpty()) {
-            log.error("Article is not valide {}", fournisseur);
-            throw new InvalidEntityException("L'article n'est pas valide", ErrorCodes.ARTICLE_NOT_VALID, errors);
+            log.error("Fournisseur is not valid {}", fournisseur);
+            throw new InvalidEntityException("Le fournisseur n'est pas valide", ErrorCodes.FOURNISSEUR_NOT_VALID, errors);
         }
 
-        return FournisseurDto.fromEntity(fournisseurRepository.save(FournisseurDto.toEntity(fournisseur)));
-
+        return FournisseurDto.fromEntity(
+                fournisseurRepository.save(
+                        FournisseurDto.toEntity(fournisseur)
+                )
+        );
     }
 
     @Override
     public FournisseurDto findById(Integer id) {
-        return null;
+        if (id == null) {
+            log.error("Fournisseur ID is NULL");
+            return null;
+        }
+
+        return fournisseurRepository.findById(id)
+                .map(FournisseurDto::fromEntity)
+                .orElseThrow(() -> new EntityNotFoundException(
+                        "Aucun Fournisseur avec l'ID = " + id + " n'a ete trouve dans la BDD",
+                        ErrorCodes.FOURNISSEUR_NOT_FOUND)
+                );
     }
 
     @Override
     public List<FournisseurDto> findAll() {
-        return null;
+        return fournisseurRepository.findAll().stream()
+                .map(FournisseurDto::fromEntity)
+                .collect(Collectors.toList());
     }
 
     @Override
     public void delete(Integer id) {
         if (id == null) {
-            log.error("Fourn isseur ID is null");
+            log.error("Fournisseur ID is null");
             return;
+        }
+        List<CommandeFournisseur> commandeFournisseurs = commandeFournisseurRepository.findAllByFournisseurId(id);
+        if (!commandeFournisseurs.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer un fournisseur qui a deja des commandes fournisseur.",
+                    ErrorCodes.FOURNISSEUR_ALREADY_IN_USE);
         }
         fournisseurRepository.deleteById(id);
     }
